requirements.txt
pytest==7.4.3
pytest-playwright==0.4.3
allure-pytest==2.13.2
pytest-xdist==3.5.0
python-dotenv==1.0.0
requests==2.31.0
